+++ changelog.md
@@ -1,3 +1,10 @@
+## 3.0.1
+- [FIX][AND] Fix loader on prefrences screen
+- [FIX][AND] Disable guests pref
+- [FIX][COR] Fetch local excluded ingredient pref on init
+- [FIX][COR] Can't add two time same ingredient to excluded list in pref
+- [FIX][COR] Select an ingredient now add it to excluded ingredient pref list
+
 ## 3.0.0
 - [FEA][COR] Toaster handler for add and like recipes
 - [FEA][COR] Singleton catalog filter service